+++ src/main/kotlin/strongdmm/event/type/service/TriggerTileItemService.kt
@@ -5,4 +5,5 @@ import strongdmm.event.Event
 
 abstract class TriggerTileItemService {
     class ChangeSelectedTileItem(body: TileItem) : Event<TileItem, Unit>(body, null)
+    class ResetSelectedTileItem() : Event<Unit, Unit>(Unit, null)
 }